requirements.txt
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
idna==2.10
numpy==1.21.0
opencv-python==4.5.2.54
openpyxl==3.0.7
pandas==1.1.5
pdfdocument==4.0.0
PDFNetPython3==9.0.0
Pillow==8.3.0
pytesseract==0.3.8
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.68
requests==2.25.1
six==1.16.0
tesseract==0.1.3
urllib3==1.26.6
